requirements.txt
python-telegram-bot==20.8
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
nest_asyncio==1.5.6
python-dotenv
Pillow==10.1.0